refactor: 메세지 페이지 컴포넌트 구조 변경
