+++ src/types/message.ts
@@ -1,0 +1,7 @@
+import { IconType } from "react-icons";
+
+export interface FilterItem {
+  id: number;
+  label: string;
+  icon: IconType;
+}
